backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.9
numpy>=1.26
opencv-python-headless>=4.9
scikit-image>=0.22
Pillow>=10.2
cairosvg>=2.7
shapely>=2.0
pydantic>=2.6
